requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
django-heroku==0.3.1
django-taggit==3.1.0
gunicorn==20.1.0
Markdown==3.4.1
Panda3D==1.10.13
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
